requirements/base.txt
# Django Core
Django==4.2.21
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0

# REST API
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0

# Authentication
djangorestframework-simplejwt==5.3.1

# Forms
django-crispy-forms==2.1
crispy-bootstrap5==2024.2

# Files & Storage
django-storages[google]==1.14.2
Pillow==10.2.0

# PDF Generation
weasyprint==61.2
reportlab==4.1.0

# Excel
openpyxl==3.1.2
pandas==2.2.0

# Async Tasks
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# Utils
django-filter==23.5
django-extensions==3.2.3
python-dateutil==2.8.2

# Security
django-csp==3.8
